moved express files into their own subfolder
made sure all commands work, server starts correctly, etc.
