+++ node-video-crime-miner/package.json
@@ -15,10 +15,10 @@
   "scripts": {
     "remakedb": "cd src/postgres/ && psql -h localhost -p 5432 -U postgres -f db.sql",
     "create": "npm run build && npm run test",
-    "start": "cd out/src/ && node ./app.js",
+    "start": "cd out/src/express-routes/ && node ./app.js",
     "cli": "cd out/src/cli/ && node ./main.js",
     "build": "tsc -p .",
-    "refresh": "tsc -p . && cd out/src/ && node ./app.js",
+    "refresh": "tsc -p . && cd out/src/express-routes/ && node ./app.js",
     "test": "jest",
     "cover": "jest --coverage",
     "lint": "eslint . --ext .ts --fix"
