+++ node-video-crime-miner/src/express-routes/envConfig.ts
@@ -1,5 +1,5 @@
 import * as dotenv from 'dotenv'
-dotenv.config({ path: '../../.env' })
+dotenv.config({ path: '../../../.env' })
 
 export default {
 }
